Revamped the checkpoint module functions to fix a glaring flaw in their design.
The problem was that I assumed that stepping through a checkpoint could be
performed in the same Monad as that associated with VisitorT, i.e. we can step
through a (VisitorT m) within the Monad m.  The problem with this is that in
our stepping environment we will want to track the current context, and to do
this we might want to use a state transformer which requires as a type
parameter the type of the state information;  this results in an infinitely
recursive type since the state information contained in m itself contains a
reference to (VisitorT m).