Added back (in a separate fn) the ability to run a visitor within its own monad.
As discussed in the previous commit this function will *not* be well-typed when
the inner monad contains a recursive reference to the VisitorT in order to
store the current context.  Thus, if you want to use a StateT to store the
context information you need to step through the visitor in a monad *outside*
of it.