Modified code to use STM instead of MVars, and wrote a worker request processor.
The reason why I switched to STM is because it simplifies synchronization; while
I probably could have gotten everything to work with IORefs and judicious use of
MVars, I realized that things were getting sufficiently complicated that there
was a chance I would miss a possibility for deadlock or inconsistency.  The
semantics of STM are such that there are never deadlocks and transactions only
commit when consistent, thus eliminating these potential sources of subtle bugs.