Added support for caching "proofs" through cacheGuard and cacheMaybe.
I decided that the most elegant way to implement this was to modify
the Cache constructor of VisitorInstruction rather than constructing
a hack using existing operations.  The way that it works is on the
first pass the argument will be evaluated and checked to be respectively
True or Just x;  if this check fails then it is equivalent to a null
and no walk will ever pass through this instruction, and if it passes
then the result will be cached and future walks will skip the test.
(In particular, cacheGuard will cache the value (), which in and of
itself has no overhead.)