Rearchitected the network worker code to eliminate extraneous threads.
The new code also has the advantage that it should be much easier
to form worker groups without creating more useless channel
dancing threads as all of the workers now tag their output with
the id they were given and they can now all share the same output
channel.

For the case of a single workers, I have created with* functions
that both take care of creating the channels and specifying a
thread tag of () and --- more importantly --- they run the given
thunk in a finally block that will automatically kill the worker
when the block is exited.