Actually, you know what? Net..Worker is pointless, so out it goes! (more below)
My goal was to try and create an abstraction over the non-network Worker
module that would make it easier to network multiple Workers together,
but I am increasingly seeing that my attempts to do this have revealed
that there really isn't much more that is worth adding to the already
existing Worker class;  a simple loop that does something like
alternating between polling between it and the messages from the Supervisor
is already good enough for the kinds of things this code would be used for.

The Supervisor, of course, is a different case because monitoring what
is going on in the network requires maintaining state.  So what I will
do is move Supervisor down one entry thus getting rid of "Network"
altogether --- for now, at least.