Fixed wart by changing VisitorStatusUpdate to VisitorProgress.
I also changed "status" to "progress" to reflect the new name
of the data structure, and dropped "update" in all places
where the value was *not* an update but just the current total
progress of the run.

This is a cosmetic change that unfortunately changes many lines
of code, but it fixes a wart that was really bothering me so I
think it was worth it.