Turns out that generating and consuming Rose trees does not mix with threading.
The problem is that as the tree is explored, *everything* is kept around because
of the pointer to the root node.  The solution is either to only keep one copy
of the tree around in each process or to generate the tree in such a way (as
will usually be the case) that the structure of the tree will not be saved as
it is explored.