Moved mant of the former typeclass methods back to their call sites.
This decision was made because this way it is easy to extend
them in such a way that makes use of all the information at
the call site rather than having to add new arguments to
the function;  arguably it also simply makes more sense for
logic that only is used in a particular place to actually
appear at that place.  Finally, there *might* be a speed-up
since there is no longer a need to construct closures to
pass to the old functions.