Modified msumBalanced to use an array.
This is faster by a factor of roughly three,
though it does introduce a space overhead in the
case where only a part of the array corresponds to
paths that will be followed (i.e., because we went
down a Path and don't care about the other choices
we could have made).  I decided that this overhead
was going to be a big deal in practice because
there are not going to be many cases where one
branch is followed but the other is not followed
later.