Now Failure also returns the progress.
I was against having this feature because the progress
should *not* be re-used if the Failure was caused by a
bug.  However, while writing the tutorial I realized that
a failure need not be due to a bug but could be due to a
resource becoming unavailable --- e.g., a database
connection being lost --- and in such cases resuming the
run after fixing the issue won't create any problems since
the fault was not in the program itself;  in fact, an end
used might be very annoyed if they lost the entire run
just because a database server needed to be restarted.